The umpires are assigning
